4 - Create basic api server
