+++ .gitignore
@@ -3,3 +3,4 @@
 backend/database/.DS_Store
 backend/database/data/
 backend/database/venv/
+backend/server/venv/
